+++ internal/goscrobble/ingress_multiscrobbler.go
@@ -4,62 +4,69 @@ import (
 	"database/sql"
 	"fmt"
 	"net"
+	"time"
 )
+
+type MultiScrobblerInput struct {
+	Artists  []string  `json:"artists"`
+	Album    string    `json:"album"`
+	Track    string    `json:"track"`
+	PlayedAt time.Time `json:"playDate"`
+	Duration string    `json:"duration"`
+}
 
 // ParseMultiScrobblerInput - Transform API data
 func ParseMultiScrobblerInput(userUUID string, data map[string]interface{}, ip net.IP, tx *sql.Tx) error {
 	// Debugging
 	fmt.Printf("%+v", data)
 
-	// if data["ItemType"] != "Audio" {
-	// 	return errors.New("Media type not audio")
-	// }
-
 	// // Safety Checks
-	// if data["Artist"] == nil {
+	// if data["artists"] == nil {
 	// 	return errors.New("Missing artist data")
 	// }
 
-	// if data["Album"] == nil {
+	// if data["album"] == nil {
 	// 	return errors.New("Missing album data")
 	// }
 
-	// if data["Name"] == nil {
+	// if data["track"] == nil {
 	// 	return errors.New("Missing track data")
 	// }
 
-	// // Insert artist if not exist
-	// artist, err := insertArtist(fmt.Sprintf("%s", data["Artist"]), fmt.Sprintf("%s", data["Provider_musicbrainzartist"]), tx)
-	// if err != nil {
-	// 	log.Printf("%+v", err)
-	// 	return errors.New("Failed to map artist")
-	// }
+	// // Insert track artists
+	// for _, artist := range data["artists"] {
+	// 	artist, err := insertArtist(artist.Name, "", artist.ID.String(), tx)
 
+	// 	if err != nil {
+	// 		log.Printf("%+v", err)
+	// 		return errors.New("Failed to map artist: " + artist.Name)
+	// 	}
+	// 	artists = append(artists, artist.Uuid)
+	// }
 	// // Insert album if not exist
 	// artists := []string{artist.Uuid}
-	// album, err := insertAlbum(fmt.Sprintf("%s", data["Album"]), fmt.Sprintf("%s", data["Provider_musicbrainzalbum"]), artists, tx)
+	// album, err := insertAlbum(fmt.Sprintf("%s", data["Album"]), fmt.Sprintf("%s", data["Provider_musicbrainzalbum"]), "", artists, tx)
 	// if err != nil {
 	// 	log.Printf("%+v", err)
 	// 	return errors.New("Failed to map album")
 	// }
 
-	// // Insert album if not exist
-	// track, err := insertTrack(fmt.Sprintf("%s", data["Name"]), fmt.Sprintf("%s", data["Provider_musicbrainztrack"]), album.Uuid, artists, tx)
+	// // Insert track if not exist
+	// length := timestampToSeconds(fmt.Sprintf("%s", data["RunTime"]))
+	// track, err := insertTrack(fmt.Sprintf("%s", data["Name"]), length, fmt.Sprintf("%s", data["Provider_musicbrainztrack"]), "", album.Uuid, artists, tx)
 	// if err != nil {
 	// 	log.Printf("%+v", err)
 	// 	return errors.New("Failed to map track")
 	// }
 
-	// // Insert album if not exist
-	// err = insertScrobble(userUUID, track.Uuid, "jellyfin", ip, tx)
+	// // Insert scrobble if not exist
+	// timestamp := time.Now()
+	// fmt.Println(timestamp)
+	// err = insertScrobble(userUUID, track.Uuid, "jellyfin", timestamp, ip, tx)
 	// if err != nil {
 	// 	log.Printf("%+v", err)
 	// 	return errors.New("Failed to map track")
 	// }
-
-	// _ = album
-	// _ = artist
-	// _ = track
 
 	// Insert track if not exist
 	return nil