+++ web/src/Pages/User.js
@@ -4,9 +4,11 @@ import './User.css';
 import { useHistory } from "react-router";
 import AuthContext from '../Contexts/AuthContext';
 import ScaleLoader from 'react-spinners/ScaleLoader';
-import { getUser } from '../Api/index'
+import { getUser, patchUser } from '../Api/index'
 import { Button } from 'reactstrap';
+
 import { spotifyConnectionRequest, spotifyDisonnectionRequest } from '../Api/index'
+import TimezoneSelect from 'react-timezone-select'
 
 const User = () => {
   const history = useHistory();
@@ -14,6 +16,11 @@ const User = () => {
   const [loading, setLoading] = useState(true);
   const [userdata, setUserdata] = useState({});
 
+  const updateTimezone = (vals) => {
+    console.log(vals)
+    setUserdata({...userdata, timezone: vals});
+    patchUser({timezone: vals.value})
+  }
 
   useEffect(() => {
     if (!user) {
@@ -45,6 +52,12 @@ const User = () => {
         Welcome {userdata.username}
       </h1>
       <p className="userBody">
+      Timezone<br/>
+      <TimezoneSelect
+          className="userDropdown"
+          value={userdata.timezone}
+          onChange={updateTimezone}
+      /><br/>
         Created At: {userdata.created_at}<br/>
         Email: {userdata.email}<br/>
         Verified: {userdata.verified ? '✓' : '✖'}<br/>